refactor: code separation and api updates
